475: funnel paths in a worker
